CGsolver and CGPoisson for CUDA and C++
I first wrote the primary code in C++ to outline the functions necessary for CUDA version and move to CUDA.
CGPoisson.cu is still developing, there is a problem implementing the operation of discretized Laplace operator. 
Note that different from the ordinary N*N size for the flattened vector x representing the \phi(x,y) field, I remove the boundary of the \phi field and implement the boundary condition in the source term b. For the mathematical details, see https://en.wikipedia.org/wiki/Discrete_Poisson_equation.
In the CUDA code, I use (1) shared memory for dot product of two vectors, (2) grid crossing for addition or substruction of an arbitrary long vector, and (3) basic cuda memory allocation.
I tried to use atomic addition for double in dot product operation, however, it doesn't work on the hardware I use.
To generate the input source file for CGPoisson.cpp and CGPoisson.cu, you need to produce a txt file storing the grid size N in first column, \rho(x,y) as a vector in the second and boundary condition (4*N) in the third (in order x=0, x=1, y=0, y=1).